Add basic API and a simplistic Espeak engine for demonstration purposes
Set up a system for having multiple engines, with only one in use at
a time (available as the Phonos.Engine DI service).

Rename the 'word' parser function argument to 'text', so the same
nomenclature can be used throughout.

Add a front-end <audio> player just as a means to test the API; it
will likely need completely re-doing.

Bug: T310140
Change-Id: I9a293004eedc259296ab04c49090056187df8908