GoogleEngine: trim unnecessary XML and remove outer slashes from IPA
Google charges by the character, so we shouldn't include anything we
don't need. The <?xml/> element is not included in Google's example
usage, and the <lang/> attribute is only used if you want to change the
language for specific words within the SSML.

We also remove the slashes around the IPA, if present, since we
apparently get better results without them.

Bug: T311233
Bug: T313497
Change-Id: I1093511f3ecc42f1bed043b57d9bc348102801e1