Introduce a 'file' param to play human-recorded audio and remove 'type'
This allows the user to use human-recorded audio instead of an IPA
rendering. For now, it uses the existing (lack of) design and simply
embeds an <audio> tag. The JavaScript introduced with this patch could
in the future be used to pull cached files of IPA renderings too,
avoiding the need to call the action=phonos API.

If the file is missing or is not playable audio, a link to the file is
returned.

This patch also removes the as-of-yet unused 'type' parameter.

The parser tests don't respect $wgUseInstantCommons (which may or may
not be set in CI anyway) so it just tests for a missing file.

Bug: T314068
Change-Id: I6bb589bd362014fdfb7ae5d56f8447c38cac906e