wgPhonosApiProxy: Add proxy support
Move the options passed to requestFactory into a variable.
Check if wgPhonosApiProxy is set.
If it is, set the proxy option to that value.
Add config variable to extension.json, and set the default value to false.

Don't copy an old extension's implementation (globals)
and instead use `$config->get( 'PhonosApiProxy' );`

Bug: T314313
Change-Id: I5111ecc322fa6b89a51d83a2ea14e9d773cc62c0