Add file caching, defaulting to FSFileBackend, and cleanup script
Introduce the Engine abstract class which all Engines should extend.
This makes the code a bit more DRY by itself, while putting the caching
methods in one place.

The storage backend defaults to FSFileBackend that goes in the upload
directory (or $wgPhonosFileBackendDirectory, if set), or the backend can
be changed entirely with $wgPhonosFileBackend.

deleteOldPhonosFiles.php is a maintenance script that can be used to
delete old files before the given date.

Performance-wise, this should probably batch all writes to the file
system on page save, but Phonos currently only works through the action
API so this can be addressed in future patches.

Bug: T314295
Change-Id: Ic916b8c7092d91a599f1c4173880789253a7838e