GoogleEngine: Normalize apostrophes and remove parentheses
This improves Google's handling of IPA. The apostrophe replacement is so
that the correct symbol (U+02C8) is used instead of the one on a
standard keyboard. The parentheses are reportedly optional to begin
with, and since Google seems to break entirely when they are used, we
simply remove them.

Bug: T313711
Change-Id: I9b996d4ea61a7b45d05e7cb14d03f62c2ded2331