Use MP3 instead of WAV
MP3 has a smaller footprint and is more widely supported than WAV.

Only the GoogleEngine provides MP3 natively. For the others, we use Lame
to convert WAV to MP3. This is installed already on WMF production.

Also change all executables to be configurable so that the extension can
be used on non-Unix-like systems.

Bug: T315119
Change-Id: I74dcc7db7c3051d41be6ebf6968df654bbb215cf