Treat phonos parser function as expensive
If the limit is exceeded, we show nothing and the page is automatically
put in a tracking category for editorial review.

The 'expensive parser function' count is only incremented if the file=
parameter isn't given, which doesn't do anything expensive.
Unfortunately we can't determine if the file is cached without making an
otherwise unnecessary trip to Shellbox, but this will change once we
start calling the Engine from the parser hook (T315481).

Bug: T315483
Change-Id: Ib227c580ccd81c8d2b6266c72df993f9cb652e6e