Always persist audio data to disk
Call the Engine from the parser hook, persisting the audio to disk and
passing that to the view. This is in line with how Extension:Score
works. With this, the API is now unused and may be removed with a later
patch pending legal decisions (since it can act as an unfettered proxy
to Google's API).

The cache key is persisted even when the page is re-parsed, meaning
there are no unnecessary API calls unless the parameters to the parser
function changed.

Bug: T315481
Change-Id: Iaad1f8f15ae0e9439b333f4995eec00b0707853d