Catch all PhonosExceptions and render localized errors to user
This ensures parsing of pages is not interrupted. Errors are passed to
the clientside and shown through the same popup we use when a file is
missing.

Since we're capturing these exceptions, they won't show up in Logstash.
Instead, we record the errors and their types into statsd so we can keep
an eye on them.

Bug: T316700
Change-Id: I5ac4db519cbf8f176db762ba253e22001de5d876