PhonosButton: show error with link to purge/reload when file is missing
After a page is parsed and loaded, it's possible the file could be
deleted (say via maintenance script or some other cache invalidation).
In this case, the client doesn't know the file is missing until the user
attempts to play it.

This patch captures any error that prevents audio playback and shows a
message prompting the user to refresh the page and try again. The link
will make a POST request to action=parse, forcing the backend to
(hopefully) recreate the file.

Bug: T316006
Change-Id: I87fbb6b10c4e0aa287700bf74a41849b70869500