Manually construct storage paths
In production (which uses Swift), the 'global-multiwrite' storage
backend enforces two hash levels used in directories. There is no
internal method to build this URL, so we have to build it manually.

In doing so, to keep parity between storage paths using the default
FSFileBackend, we set the containerPaths which effectively tells
FileBackend to ignore the domainId. This is not normally what we want
but should be safe for Phonos since files with identical parameters
(including language) won't differ cross-wiki and should be shared. All
storage paths are still in a dedicated 'phonos' directory so there
should be no collision with other backends using the same storage
system.

This is the same system Extension:Score uses and should make Phonos
usable in production.

Also rename some methods for consistency.

Bug: T317417
Follow-Up: Iaad1f8f15ae0e9439b333f4995eec00b0707853d
Change-Id: I56d7554ae60a662a4742b1e0ba7f0fbba27a2ef9