Show error if lang not supported
Check whether the user-supplied langauge is supported by the
engine, and throw an error if it's not.

If an engine doesn't report a list of supported languages
then they are assumed to support all languages and no error
is ever shown. Only Google and Espeak are done in this patch.

Language codes are normalized to some extent (lower cased
and underscores replaced).

Bug: T320523
Change-Id: I0858c2aaf9cf879413fe74f7c4ea55ea0aeb3d91