Add a minimum file size and show an error if a file less than it
When invalid IPA is provided or if the engine fails to interpret it, a
small, blank file is often still created. Editors may do trial and error
before the file is actually generated, all the while creating unused
orphaned files. This commit adds a minimum file size to guard against
this.

Ideally we'd instead look at the duration of the generated audio, but
this requires a round trip through Shellbox which we can seemingly avoid
by solely looking at the file size.

eSpeak and other engines may generate smaller files, so for now we only
set the minimum for the Google engine.

Bug: T324239
Change-Id: I11a471e5fc6a2e5884e460c6ceb2d09793953ec0