GoogleEngine: don't remove parentheses from IPA input
After T324239, empty files created by Google (such as unsupported IPA)
now show an error. Parentheses have a special meaning that we should
instead be upfront about not supporting.

There is other normalization we're still doing for Google that may be
removed later, but it seems certain parentheses should not be removed.

Bug: T323912
Change-Id: I0129949269fedadd559e1e793aeddb99f78ddf2b