Improve icon size and position, and baseline of label
Change all relevant elements to be inline, with baseline alignment
so that their text aligns with the surrounding text. The icon is
dropped 0.1em so that it appears more centered.

A new 'emptylabel' class is added so that the label can be hidden
when not needed, to avoid its margins interferring with the
layout.

This change also means that layout works better in RTL direction,
with correct spacing in all cases. Note that this follows content
language and not interface language, because Phonos is content.

Bug: T320820
Bug: T320415
Change-Id: I047453f0fd874029057cc659195176d71653692d