Store usage of Phonos files as page properties
This allows us to more easily identify used files, and in turn, *unused*
files via the countOrhpanedFiles.php script, which has been been updated
in this commit.

The page property is set in LinksUpdate after an edit. If Phonos doesn't
create the file immediately and instead queues a job, the page prop is
still set. Although unlikely, it's possible the job will fail and we
have a recorded usage of a file that doesn't exist, but the maintenance
script doesn't care and the page props will get refreshed eventually
anyway.

You can force the page props to be built with:

  maintenance/run refreshLinks --category='Pages that use Phonos'

Bug: T326163
Bug: T324233
Change-Id: I3299ab26854c4facedeade4524b06108f3592cb3