Revert PhonosButton to display:inline-block instead of inline
We changed the outer PhonosButton element to be display inline
instead of inline-block (OOUI's default) in order to resolve
baseline issues, but we've since sorted those issues on the
inner elements and so no longer need to change the outer one.

This fixes a bug with RTL scripts and icon layout.

Bug: T320415
Change-Id: I2a6f276697f45e3d5fd1bb0d5673516147c09a0c