Avoid linking to invalid filenames in error message
The getErrorMessage() method is used to (among other things) tell
the user when a filename is incorrect. If the filename is malformed
then mw.Title() throws an exception, but we don't catch it.

This change catches this exception and falls back to just showing
the plain text as supplied in the error message.

Bug: T345672
Change-Id: I21923be7ba6e5d6cb5ed73c8966c98efd5003bcb