Delay loading ext.phonos module until user clicks
In order to reduce the size of the initially loaded modules,
create a new small ext.phonos.init module that registers a click
handler on Phonos elements. Clicking then loads the larger
ext.phonos module, and infuses the OOUI button.

Also change the dependency of ext.phonos to oojs-ui-core rather
than oojs-ui-widgets, because the required PopupButtonWidget
and PendingElement are in core.

Bug: T345414
Change-Id: I69cb817386c5865d697e2a216cc72437c31e3d6a