Add the Phonos init module as a dependency of the main module
If a page used Phonos before I69cb817386c58, it would have ext.phonos
in its module list, saved into the parser cache with a long expiry
time. After that change, ext.phonos no longer attaches events to the
buttons, so clicking the button on a cached page navigates away to the
file.

So, add ext.phonos.init as a dependency of ext.phonos, so that when the
cached page requests ext.phonos, events are still attached.

Bug: T345414
Change-Id: I3d46c1a74ff07e74cce07327c9793e344699573e