Trigger button on enter as well as click
Add a keydown listener to call the button init handler when the
button element is pressed via the enter key (i.e. when it has focus).
This duplicates what OOUI already does, but we don't want to load
that here.

Note that the element that has focus is the <a>, which is inside the
ButtonWidget.

Bug: T347056
Change-Id: Icd170018f98365b367a6e5bad51e61ecc40633a0