Use static integers instead of OO.ui.Keys
OO.ui.Keys is defined in the module 'oojs-ui-core', but a dependency on
this module is missing. Therefor the access of this object can fail if
the module is not yet loaded.

This change fixes this problem not by adding a dependency on the module
because this module is very large. This change use static integers for
these ASCII characters.

Bug: T350369
Change-Id: I10c7fd827be559ba8dcdec01c38ee6ace63922aa