Generate error messages on client side
The change allows to show the error messages in user interface
language, because only the message key and not the expanded message
is in the rendered and cached HTML.

The JSON in the HTML now uses an array instead of a string in "error"
to distinquish between the new and the old format. The old format is
still supported by the JavaScript module. This can be removed after the
cache has been expired by I0b68ea9f39458921fa368fd3a68581e41893034f.

The missing file link in the popup is now blue instead of red and has
no &action=edit&redlink=1 anymore. This will be changed in
I0d268d755364e9e12ecbca214590bc755917f56e.

The method PhonosException::toString() is not needed anymore.

Bug: T347682
Change-Id: I4109a5bae4f42ad3b3767864473e49f0a2bdb677